errumamos venda

+++ Components/Models/VendaProduto.cs
@@ -6,8 +6,8 @@
         public int IdVenda { get; set; }
         public string Produto { get; set; }
         public int Quantidade { get; set; }
-        public decimal ValorUnitario { get; set; }
-        public decimal ValorTotal { get; set; }
+        public double ValorUnitario { get; set; }
+        public double ValorTotal { get; set; }
         public bool Status { get; set; }
     }
